requirements.txt
langchain-google-genai
google-generativeai
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-chroma
langchain-text-splitters
pydantic
pytesseract
pillow
pymupdf
fitz
chromadb
python-dotenv
